requirements.txt
customtkinter
reportlab
pytest
pytest-cov
pyserial
